backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
Flask-SocketIO==5.3.6
python-dotenv==1.1.1
pyserial==3.5
requests==2.32.3
